refactor(actions): use new pattern to get state
zustand's set cannot be called async (see
https://github.com/pmndrs/zustand/discussions/863#discussioncomment-2352736)
and the elkjs layout algo is async, so we cannot use the pattern
of determining changes within a function passed to the set method.
also, by mutating a deep-copied result of getState instead of
setState's argument, we no longer need immer as middleware.