feat: use elkjs for layout instead of dagrejs
when two layers have many connections, dagrejs spreads out the nodes
a lot (see #112 description). the elkjs layout does not do this, and it
provides significantly more customization options.
also set the layout option to preserve order if layout is good enough.
annoyingly, elkjs layout algo is async, so had to adjust a couple
invokers to be async.