touchup: elevate edge svg based on selected parts
when we were using transparency for background edges,
the fact that the edge svg would elevate all edge paths
in front of background nodes and edges was not noticeable.
since we turned off transparency, this caused this elevation
to be noticeable and confusing.

now we are only elevating the edge svg for selected or
neighboring edges, and other edges do not have their svg elevated.