fix: topic only saves if redux extension installed
before this commit, topic data was only being saved
if the client's browser had redux devtools installed.

this was an absurd issue, but ultimately zustand has a `setState`
method, which we rely on because of our "no store actions" pattern,
but the zustand middleware example only shows overriding the `set`
method, so we only overrode that one.

the `set` method override would be sufficient if the client has redux
devtools installed because zustand's devtools middleware invokes `set`
when `setState` is called.

we definitely don't want to rely on this, so we just need to override
`setState` as well, to ensure our middleware is always triggered.