chore(deps): remove renovate
when installing renovate, there were a lot of
things to upgrade, and some upgrades required deep investigation, and
their solutions were not obvious.

this seems somewhat likely to happen whenever upgrades
are made, and it doesn't really seem worth addressing those
unless the upgrades are actually adding value.
so we're not going to do automatic upgrades.