refactor: check playground by topic.id undefined
instead of topic being null. this is needed so that
the playground can use topic.description even
though it will have no topic saved to db.

note that this commit breaks the store until migration
is added, but want to keep migration in the next commit
because it includes description migration.