fix: allow edge custom label to be reset to null
this was preventing "change edge type" from working with an edge
that has a custom label because the type would change but the custom
label would remain.
needed to ensure custom label could be null instead of undefined,
since prisma understands `undefined` as not to be updated for updates.

also made so that custom label does not get set if it's the same as the
edge type.