feat: add obstacle node
intended to convey blockers to a solution, but aren't as hard-stop as
"blocker".
critique can be used as well, but this is a little more specific,
is something more preventative than detriment-like.

particularly didn't want to use "problem" because problem is used in a
filters as a central thing to solve; concerns likely aren't a central
focus of the topic.

e.g. "road is a thoroughfare" is obstacle of "speed bump"
e.g. "where?" or "no place to put it" is a concern of "dunk tank"