touchup: default problem filter to all details
with a large diagram with multiple problems, this seems preferrable,
with a small diagram with only one problem there's probably
preference for defaulting to not show criteria/solutions,
but it doesn't seem worth covering this case.

also, the general node type filter can remove solutions/criteria if
users prefer that default.