touchup: add option for forceShownIndicator
not sure if it's best to have toggle per indicator like this,
but we'll keep an eye on the UX of it.
it's nice to have off for reducing clutter and just looking at the
diagram, but nice to have on when modifying the view, to easily
understand the filtering.