refactor: getSelectedId -> getSelected
some methods (e.g. deleteGraphPart) are easier with the graph part
itself, rather than just the id.

doing this should make it easier to add e.g. hotkeys for deleting graph
parts.

could've alternatively changed deleteGraphPart to accept id
instead of part, but it's hard to see much benefit of that over this.