fix: diagram screenshot doesn't zoom out to fit
if diagram was big, and wouldn't fit in the resolution, the diagram
wouldn't fit in the screenshot.

the minimum zoom was just not zooming out far enough.
we might as well set the minZoom really low so that low-res screenshots
can still fit a bunch of nodes if people want them to.