touchup: don't layer effects apart from components
it seems fine for a solution's direct effect to be in the same layer as
a solution's component.

this makes it easier to distinguish solution direct effects from
component effects.

also order components to the left of effects because they seem
more important/core to solutions, so should come first when
reading left-to-right.