fix: store syncing unintentionally
previously would always sync when store changes, except when topic.id
doesn't change.

this was a problem because an SPA navigation that pulled in new changes
would treat those as a diff and try to save them again.

now, we add a flag to syncing middleware to control syncing - off when
we're pulling data from local storage or the API, on otherwise.

note: we also don't sync if we're on the playground.

note2: this logic is being duplicated across stores. didn't want to
take efforts to make code reusable, since this solution might not be
used long-term (could switch to CRDT later).