touchup: don't show indicators on hover/select
because it seems jarring when you hover
something and it reveals something that's immediately hovered,
e.g. hover edge -> hovering score.

keeping the change to use css instead of js for show/hiding
because it seems a little cleaner; could consider js for
performance improvements in the future (because it could
prevent running rendering logic).

also keeping diagram-node/-edge classes because they seem probably
useful in the future.