docs: add "show indicators" to tutorial & docs
This was annoying to do because of the reasons below, but since we're
defaulting indicators to hidden, we have to ensure docs mention it
(since users would otherwise be confused when getting to "scores" in
the docs).

I think the button name is slightly suboptimal because:
- Score is more interactive than just an Indicator

Also these docs updates are a little suboptimal because:
- the button is only explained in the Scoring section; hopefully
users can understand its use for the indicator section, but it seems
annoying to duplicate the explanation there too.

It's possible that "show scores" should be a separate button, but:
- both would often be on/off together
- both would need separate tutorial/docs mentions?