touchup: allow node click to select or change text
regardless of context, so now you can edit text without selecting
the node, and select a node in the details pane without the details
indicator button.

We set `cursor: pointer` to convey that a node will be selected vs
`cursor: text` to convey that node's text will be selected. This
seems good enough to avoid confusion, at least on desktop... We'll
want to keep an eye out for if this is acceptable on mobile.

also includes setting role button for diagram parts
so they can be clicked by vimium.