fix: can save view with no order
`-Infinity` was passed over the API if all views were deleted, and a new
one created.

Unfortunately Prisma serializes `-Infinity` as null
(https://github.com/prisma/prisma/issues/19966#issuecomment-1658086903),
so this API request was able to save as null in the db, rather than
erroring as out of integer bounds.

Added `.finite()` to the zod schema to prevent this possibility in the
future.

No other `number`s in schema allow null, so this is currently the only
field that can have this issue, but we'll need to keep an eye out for
it. `int`s that are conditionally `null` are susceptable.