chore: remove storybook
wasn't really using this, it's just being an extra dependency.

if we want to use storybook, we'll want to start from scratch anyway,
since the current usage is outdated.

includes removing legacy-peer-deps config because it was added for
storybook. emotion and eslint happened to be depending on this too,
so those versions were loosened to not require it anymore.