fix: headline appears on three lines,
not sure why on desktop it seemed to be 2 lines, but on laptop
and on ci it was on 3 lines.

this was consistent for each screen size too e.g. small & large when
on desktop.

maybe browser version renders differently? in any case, increasing
headline container width should fix this.