touchup: "details" -> "notes" indicator
don't need an indicator to be able to select a node anymore, because
clicking node border allows this.

so the details indicator doesn't need to always show, and can just be
treated as a content indicator for notes.

not enough space in the content indictor section though, so just
shoehorned it into the status indicator section for now.

not ideal but it works. really wanted to do this so that UI can be less
cluttered when showing indicators.