refactor: extract selectedGraphPart to new store
this property often didn't want to be with the view

e.g. viewX (like viewJustification) want to preserve the selected node,
while changing the rest of the view state.

e.g. selecting/unselected node should not be a view change,
should not trigger "changed filter" event, should not be undone/redone
when using that navigation buttons...

the main use case was to allow saving selection with a quick view.
short-term we can add ?selected to URL for this, long-term we can
add `selectedState` to `views` table, with a "save selected state?"
option (similar to what we'll want for perspectives).