touchup: add New Topic to site header
nice to not need to go to My Topics first to create a topic.

could make this open the Topic Form directly, but /new allows opening in
a new tab, which seems desirable e.g. if editing another topic.

also removed the icon from the New [topic] button in the topic list,
since the site drawer's New Topic makes more sense with a plus button,
and the topic list button looks a little off with a plus button.