touchup: don't update URL or viewstate on recalc
i.e. when filter changes, we should just calculate if a view is
selected, we don't need to set new view state or update the URL.

without this, URL would update automatically, which is a bit annoying.

also without this, sometimes the view would refit unintentionally
(because a view matched to be reselected).