touchup: make details section headers stand out
because it was hard to see them previously. root problem seems to be
that each details section has high saturation and can't much rely on
standard indentation because of nodes/add-node buttons being there.
considered adding transparency for each section's content until it's
hovered/focused, but that also kind of makes it seem like it's disabled,
which is a little awkward.

mainly increased font size, bolded, centered.

also:
- generally replaced list items with divs because those weren't great
semantically
- added sections around each section because that makes sense
semantically