touchup: show scores by default
they don't seem to add _too_ much clutter, and scores are important
enough that it seems ok to default on.

new users don't find them very easily when they're off.

maybe we could also try making it more obvious how to turn scores on,
or add a tooltip or something. but this is good enough for now.