feat: add disagreement scores
includes:
- add aggregation modes
- choose score colors by score meaning
- extract custom NestedMenuItem for consistent styling

considered using H/M/L instead of numbers for disagreement, so that
they stand out clearly from regular scores, but those can't be trivially
used with solution totals, so it seemed better to be consistent with
using the same score numbers. potentially the score shape can
distinguish disagreement in the future.