touchup: only show core node outline in diagram
seemed like a ton of extra clutter in the core node summary.

can consider establishing the outline-to-indicate via usage elsewhere
that doesn't add too much clutter. maybe via core node icons
that'll show in the corner of most pages?

can still consider removing the outline even in the diagram, but we'll
continue seeing how that feels for now.