touchup: show node toolbar on focus, not hover
seems possibly worthwhile not to show on hover:
1. it can be overwhelming to users when there's a lot of flickering as
the mouse is dragged across the screen
2. toolbar can be accessed via right click if desktop users want to be
faster (well 'see in summary' isn't there yet but it could be)
3. seeing multiple toolbars at once seems completely unnecessary

the reason for showing on focus is: without on-hover, mobile can no
longer tap into a node's text to open the toolbar, but that's nice to
have, so focus helps with that.

tabIndex added to the node because without it, if text is focused and
then the node is selected, the toolbar flickers off and then back on
because the node itself couldn't hold focus.
not sure if this is suboptimal because now the node and its textbox are
separate tab targets, but we can remove that if it seems particularly
undesirable.