touchup: show edge label border on neighbor edges
this was originally on then turned off because of clutter when
there are several neighbor edges.

this does add clutter but it looks awkward when a neighbor edge
is elevated above a node and no border is added to the label,
because then the label is hard to distinguish from the node.

not sure what the ideal strategy is here, will have to keep
thinking on it.