refactor: align column addables with found nodes
i.e. add button should show same relations to add as the relations
that are used for finding nodes to put in the column

did this by adding `directedSearchRelations` for each aspect filter,
and filtering the default addables for a summary node via this.

main motivation for this was for making it easier for align
these things with nodes that can be found via add node menu search.

also this made it much easier to show the summary add buttons based
on the direction they need to add in (e.g. effects use 'createdBy'
if problem effect, 'creates' if solution effect). (which is now being
done properly - before this refactor, it wasn't).

includes:
- adding header and add buttons to incoming/outgoing columns (same as
add buttons in diagram)