touchup: add tooltip to score button
mainly because when quick scoring is off, it looks like you can click
the score like a button, yet there's no indication of what will happen
on click.

includes:
- small score button refactor to forward ref so that mui tooltip can
wrap it
- removing modal-ification (darkening the background) of score pie
when score is clicked, because that doesn't seem totally necessary,
and can be annoying to flicker the screen when scoring a few different
nodes.