touchup: try only expanding common add buttons
this allows the advantage of quickly clicking common buttons, without
the cost of an overwhelming number of quick options.

unfortunately this button group is starting to get very complex,
but I'm hoping we can find a UX path that gets the best of both worlds,
so we can remove the option to expand these buttons altogether.