touchup: combine above/below add buttons
seems more confusing for users if they're separated - this way they
don't have to know which direction nodes will lay out in.

motivated by wanting to deprecate / not maintain logic for
determining which nodes specifically are added above/below.

includes:
- showing all uncommon addable relations from a node, rather
than just the first one found for the to node type; mainly
this was an issue because "problem has problem" and
"problem causes problem" now need to both show up in the
combined add menu.