refactor: point edges directly at R.Flow handles
instead of pointing at nodes and letting React Flow calculate
which handle to use based on source/target handles.

motivated because we're about to flip edge direction, and
reactflow by default assumes handles can only exclusively be used
as source OR target. without this refactor, if an edge's direction
is flipped, actions like reconnecting will be started by
dragging from the wrong, opposite handle... reactflow does this
UNLESS the handle id's are used by the edges.

so now during layout we determine which port the edge should use,
then render handles on the node based on the ports used, and
store port ids on edges (elk port = react flow handle).

we're rendering handles on the node based on ports used because
we likely will use a second port (on each side of a node) in
the near future in order to have incoming edges visually separated
from outgoing edges.

includes:
- a FIX for showing handles when connecting, since part of that
logic was accidentally added in a previous commit without the
rest of the logic to support it