fix: add node button's tooltip "invalid anchor"
would happen when:
1. the flow node is hovered, displaying the add node button,
2. the add node button is hovered, displaying the mui tooltip, and
3. the button is unhovered, hiding the add node button

I believe this would happen because the mui tooltip expects the
add node button to be in the DOM in order to know how to
position itself.

this fix relies on using "invisible" instead of "hidden", so the add
button is still in the DOM. this doesn't really fix any functionality,
but the current functionality is acceptable. hiding the warning is
what's important because it gets thrown so often that it's hard to see
potentially-important warnings otherwise.