touchup: transition smoothly on quick view change
I think we originally didn't do this because a lot could be changing at
the same time? i.e. viewport + node animations.
but increasing the duration of the transition seems to make it ok, and
having a transition makes the view change feel smoother.

Ideally we probably would layout starting from origin in the center of
the diagram, so that a view with only a couple nodes wouldn't be
completely in the corner when compared to a view with a bunch of nodes.