refactor: set positionedDiagram right after layout
rather than waiting until a render _after_ the layout.

I think this may have done in the past in order to show a loading state?
but we can always add a `loadingLayout` state if we need to, and that'd
be more explicit anyway.

also moved `selected` setting into Diagram because that makes more sense
to do at the last moment.